Include handling of coverage and quality report date being handled by transformAboutContent. Uses the most recent date when the coverage-and-quality-report.html was updated
